Additional Edits to Fix Merging
