+++ webpack.config.js
@@ -4,7 +4,7 @@ const CopyWebpackPlugin = require('copy-webpack-plugin');
 module.exports = {
   entry: {
     script: './src/Index.tsx',
-    content: './src/Content.ts',
+    content: './src/Content.tsx',
   },
   output: {
     path: path.resolve(__dirname, './dist'),
